Updated setting scene

+++ IK_game/src/Button/Button_settings_audio.java
@@ -7,7 +7,7 @@ import Scene.GlobalScene;
 public class Button_settings_audio extends GlobalButton
 {
 
-	public Button_settings_audio(int x, int y, int w, int h, DrawbleObject parent)
+	public Button_settings_audio(double x, double y, double w, double h, DrawbleObject parent)
 	{
 		super(x, y, w, h, parent);
 		this.setText("Audio");
