+++ IK_game/src/Button/Button_settings_video.java
@@ -11,7 +11,7 @@ import Scene.GlobalScene;
 
 public class Button_settings_video extends GlobalButton
 {
-	public Button_settings_video(int x, int y, int w, int h, DrawbleObject parent)
+	public Button_settings_video(double x, double y, double w, double h, DrawbleObject parent)
 	{
 		super(x, y, w, h, parent);
 		this.setText("Video");
